refactor : 감정 우표 목록에 ConstraintLayout 수직 체인(spread) 연결 및 가이드 라인 개수 축소
