--- .github/ISSUE_TEMPLATE/mission_template.md
@@ -1,18 +1,0 @@
----
-name: "🚩 미션"
-about: "미션 단위로 관리하는 이슈 템플릿"
-title: "[Mission]"
-labels: []
-assignees: []
----
-
-## 🚩 미션 설명
-<!-- 이번 미션의 요구 사항과 구현 목표를 간단히 작성해주세요. -->
-
-## ✅ 체크리스트
-- [ ] 
-- [ ]
-- [ ]
-
-## 💬 참고 자료
-<!-- 관련 문서, 링크, 이미지 등이 있다면 첨부해주세요. -->
