--- .github/PULL_REQUEST_TEMPLATE.md
@@ -1,15 +1,0 @@
-## 🚩 관련 이슈
-<!-- 이 PR이 연결된 미션 이슈 번호를 적어주세요. -->
-- close #이슈번호
-
-## 📌 구현 결과
-<!-- 구현한 기능을 요약하거나, 필요한 경우 스크린샷/영상/GIF를 첨부해주세요. -->
-
-## ❓ 리뷰 요청
-<!-- 코드 리뷰어나 팀원들에게 피드백받고 싶은 부분을 작성해주세요. -->
-
-## 🤔 질문
-<!-- 구현하면서 궁금했던 점이나, 추가 논의가 필요하다고 생각되는 부분을 적어주세요. -->
-
-## 💬 기타 공유 사항
-<!-- 그 외 추가로 공유하고 싶은 내용이나 필요한 정보를 작성해주세요. -->
