+++ app/src/main/res/drawable/group_573_1.xml
